iter2: Sum in double precision

+++ internal/api/handler/balance/get/response.go
@@ -1,6 +1,6 @@
 package get
 
 type Response struct {
-	Current   float32 `json:"current"`
-	Withdrawn float32 `json:"withdrawn"`
+	Current   float64 `json:"current"`
+	Withdrawn float64 `json:"withdrawn"`
 }
