+++ internal/api/handler/withdrawals/response.go
@@ -6,6 +6,6 @@ type Response []Withdraw
 
 type Withdraw struct {
 	Order       string       `json:"order"`
-	Sum         float32      `json:"sum"`
+	Sum         float64      `json:"sum"`
 	ProcessedAt api.Datetime `json:"processed_at"`
 }
